refactor(observability): 归位 Trace 适配与分段能力
