+++ src/observability/skill-health-analyzer.ts
@@ -23,8 +23,8 @@ import {
   type CcSession,
   type TraceSession,
   type SkillSegment,
-} from './trace-adapter.js';
-import { legacyCcSessionToTraceSession } from './trace-source.js';
+} from './trace/index.js';
+import { legacyCcSessionToTraceSession } from './trace/source.js';
 import { createTraceSessionIndex } from './trace/session-index.js';
 import { setOwnRecordValue, sumRecordCounts } from '../shared/record-count.js';
 import { checkedSumTokenCounts } from '../shared/token-usage.js';
