+++ src/observability/trace/adapters/codex/exec-command.ts
