+++ src/observability/trace/adapters/codex/protocol.ts
