+++ test/observability/conversation-catalog.test.ts
@@ -12,7 +12,7 @@ import {
   normalizeCodexRolloutIndex,
   readCodexTaskRecords,
   synchronizeCurrentCodexRolloutIndex,
-} from '../../src/observability/codex-conversation-index.js';
+} from '../../src/observability/trace/adapters/codex/conversation-index.js';
 import { createCodexConversationCatalog } from '../../src/observability/conversation-catalog.js';
 
 const roots: string[] = [];
