+++ test/observability/source-record-archive.test.ts
@@ -12,7 +12,7 @@ import {
   loadObservationSourceRecordArchive,
   writeObservationSourceRecordArchives,
 } from '../../src/observability/source-record-archive.js';
-import { forEachNonEmptyUtf8Line } from '../../src/observability/trace-source.js';
+import { forEachNonEmptyUtf8Line } from '../../src/observability/trace/source.js';
 import type { ObservationSourceRecordArchiveRef } from '../../src/observability/contracts/inbox.js';
 
 function jsonl(records: unknown[]): string {
